Load the nested models of eager loaded models (#119)
* Everyone loves RSpec's focus

* Load nested models for eager loaded models

Generated nested models register themselves upon being created by
Avromatic::Model::Builder. If Rails's autoloader didn't recreate a model that
should be eager loaded, then its nested models won't be registered as we hook on
Railtie #to_prepare that itself triggers a clean-up of the nested models
registry.

Therefore, we have to go through the attribute definitions of each eager loaded
model and try to register its nested models.

* Visit all levels of the tree

As we find an instance of an `Avromatic::Model::Types::RecordType`, we might
want to continue the search.

* Make StringType inherit from AbstractType

This was somehow missing.

* Replace case statement with polymorphism for the greater good

* No need to use map here

* Add missing test model

* Rename #referenced_models to #referenced_model_classes

Thanks for the suggestion Joel!

* Bump version and update the changelog

* Revert wrap

* Freeze things

For the sake of consistency.

* Disregard maps' key type

They can only be strings.